// dashboard/src/utils/hasura-api/generated/schemas/actionRelationship.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { ActionRelationshipFieldMapping } from './actionRelationshipFieldMapping';
import type { ActionRelationshipType } from './actionRelationshipType';
import type { QualifiedTable } from './qualifiedTable';

/**
 * Relationship from an action's output object type to a table.
 */
export interface ActionRelationship {
  /** Name of the relationship, shouldn't conflict with existing field names */
  name: string;
  /** Type of the relationship */
  type: ActionRelationshipType;
  /** Name of the source database of the remote table */
  source?: string;
  /** The table to which the relationship is defined */
  remote_table: QualifiedTable;
  /** Mapping of object type field names to the remote table's columns */
  field_mapping: ActionRelationshipFieldMapping;
}