// dashboard/src/utils/hasura-api/generated/schemas/customTypeObjectField.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */

export interface CustomTypeObjectField {
  /** Name of the field */
  name: string;
  /** GraphQL type of the field */
  type: string;
  /** Description of the field */
  description?: string;
}