// dashboard/src/utils/hasura-api/generated/schemas/enumValueDefinition.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */

export interface EnumValueDefinition {
  /** Value of the enum */
  value: string;
  /** Description of the enum value */
  description?: string;
  /** If set to true, the enum value is marked as deprecated */
  is_deprecated?: boolean;
}